FEBio2: Added the FE_PARAM_STD_VECTOR_DOUBLE parameter type. Added a FEDataModel class, which is kind of like the DataStore, but without the file interface, and managed by FEModel. Added support for new data section in Output that reads FEDataModel classes.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/FEBio2@9762 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ FEBioMech/FERemodelingElasticDomain.cpp
@@ -57,7 +57,7 @@ bool FERemodelingElasticDomain::Initialize()
 void FERemodelingElasticDomain::StiffnessMatrix(FESolver* psolver)
 {
 	// repeat over all solid elements
-	int NE = m_Elem.size();
+	int NE = (int)m_Elem.size();
 
 	// I only need this for the element density stiffness
 	FEModel& fem = psolver->GetFEModel();
